fix: renamed axios hook for better compatibility
